2/28/25 piechart image generated
we r almost done. we got 2 problems.

I can run python file in php and everything works except the image, it's not being updated or generated. this should help: https://stackoverflow.com/questions/66009499/running-python-script-in-php-script

I am not displaying tire to manager when he is managing tasks. right now, it only shows 0 tires, but it should show all the bad tires. just increment it.